All room tests for migration from 8 to 10. Refactor: moved migration definitions to their own file.
